Fixed object motion blur not respecting exposure mode setting.
Introduced new intraframe_time function to Camera class, which takes the fractional intra-frame time (in [0, 1]) and returns the exposure mode adjusted time.

For example, given FPS = 24, exposure mode is degrees, start degree = 0, end degree = 180 (ie half-frame exposure), then for input 0.6 the output is 0.025. Note that this time is after the shutter closes, which is at ~0.02083, since we specified a half-frame exposure.